+++ elsim/similarity/libsimilarity/similarity.h
@@ -36,7 +36,7 @@ typedef struct libsimilarity libsimilarity_t;
 
 #ifdef __cplusplus
 extern "C" {                                                                                                                                                                                     
-    float entropy(void *, size_t);
+    double entropy(void *, size_t);
     void set_compress_type(int);
     int ncd(int, libsimilarity_t *);
 }
@@ -46,7 +46,7 @@ size_t compress(int, void *, size_t);
 int ncd(int, libsimilarity_t *);
 int ncs(int, libsimilarity_t *);
 int cmid(int, libsimilarity_t *);
-float entropy(void *, size_t);
+double entropy(void *, size_t);
 #endif
